feat(seo): full SEO sweep across public pages — SSR metadata, JSON-LD, sitemap, robots
- Public detail pages (`/projetos/[id]`, `/entidade/[slug]`, `/usuarios/[slug]`) are now server-rendered with `generateMetadata` (title, description, OG/Twitter card, canonical, keywords) and JSON-LD (`CreativeWork`, `Organization`, `Person`). Interactive bodies moved to sibling client components; React Query is hydrated via `initialData` where the path is public.
- `/guias/[[...parts]]` resolves the URL on the server to emit per-page titles built from the breadcrumb (subsecao · secao · guia · Aquário). Unknown sections render `noindex`.
- `/sobre`, `/recursos`, and the landing `/` get per-page `metadata`.
- Static client listings (`/projetos`, `/entidades`, `/calendario`, `/calendario-academico`, `/mapas`, `/grades-curriculares`) get a sibling `layout.tsx` with metadata so crawlers and social cards see real titles in the static HTML.
- New `src/app/sitemap.ts` (revalidates hourly) listing all public pages plus every PUBLICADO projeto, entidade, and guia/secao URL.
- New `src/app/robots.ts` allowing public routes and disallowing API, admin, auth, edit/create flows, and internal pages.

Drafts/archived projetos and facade users are correctly excluded from indexing.
